Format report prompts without using FormattedHTMLMessage
